+++ CHANGELOG.md
@@ -5,6 +5,9 @@
         - added confirmation dialog
         - updated Did You Know
         - actually remembered to commit the new ReadingProgressBar class :-(
+    - plugins
+        - fixed imports
+        - fixed overrides (e.g. Filters.SavitzkyGolay)
 - 2024-10-16 4.1.14
     - AutoRaman
         - fixed "Stop" button